requirements.txt
boxsdk
boxsdk[jwt]
pymysql
requests_toolbelt